enhanced compiler checks and related fixes  / memory issue corrections / general code improvements (#195)
* added extra, common C and C++ compile flags from pelz Makefile

* removed redundant declaration for set_applog_severity_threshold()

* fixed some signed/unsigned integer conversion warnings

* fixed size_t to uin32_t cast warnings in kmyth_sgx_test.c

* whitespace changes in kmyth_sgx_test.c

* fixed missing free() in kmyth_sgx_test.c

* fixed warnings for casting const delim strings to char * (and whitespace changes)

* modified enc_get_unsealed_size() ECALL

* updated Makefile based on Intel Sample Code, fixed size_t to uint32_T cast warning

* fixed declaration is not a prototype warning

* cleaned up C vs C++ compilation

* fixed a lot of compiler warnings

* whtespace changes

* consolidated '#include ENCLAVE_HEADER_XXX' in kmyth_enclave_XXX.h

* mods resulting from valgrind analysis of kmyth-test binary

* fixed introduced cipher bugs

* added valgrind as a dependency and uncommented valgrind API call

* added checks on malloc() result and input params for demo_kmip_server_receive_get_key_request()

* edl whitespace and comment fixes

* whitespace fix in sgx/trusted/src/ecall/kmyth_enclave_unseal.cpp

* removed unhelpful log message

* added pointer parameter checks to kmyth_encrypt_data()

* updated kmyth_unseal_into_enclave() buffer size checks

* fixed SAPI/TCTI context finalization

* removed checks for non-null pointer before free() and some other one-line if statements

* added one more missing free() in cipher_test.c

+++ sgx/demo/src/util/demo_tls_util.c
@@ -88,7 +88,7 @@ void tls_get_verify_error(TLSPeer * tlsconn)
     return;
   }
 
-  int verify_result = SSL_get_verify_result(ssl);
+  long verify_result = SSL_get_verify_result(ssl);
   if (verify_result != X509_V_OK)
   {
     kmyth_log(LOG_ERR, "SSL_get_verify_result: %s",
@@ -267,7 +267,7 @@ int demo_tls_config_client_connect(TLSPeer * tls_clnt)
   //   - expected to be <IP address or hostname string>.<name string>
   //     (e.g., localhost.demoServer, 127.0.0.1.server, ...)
   char * server_name = NULL;
-  unsigned int server_name_size = strlen(tls_clnt->remote_server) + 1;
+  size_t server_name_size = strlen(tls_clnt->remote_server) + 1;
   if (tls_clnt->remote_server_func != NULL)
   {
     server_name_size += (strlen(tls_clnt->remote_server_func) + 1);
@@ -295,7 +295,7 @@ int demo_tls_config_client_connect(TLSPeer * tls_clnt)
     free(server_name);
     return -1;
   }
-  if (bytes_needed > server_name_size)
+  if ((size_t) bytes_needed > server_name_size)
   {
     kmyth_log(LOG_ERR, "truncated server certificate verification name");
     free(server_name);
